uart: Add a "txdone" register, set compat to "sifive,uart1"
The UART is a pretty small device, so its clock tends to be coupled to
at least some other part of the SOC.  As a result, users may want to
change the UART's clock frequency in close proximity to transmitting a
character out of the UART.  The UART currently does not actually expose
anything that software can read in order to check to see if transmission
is done -- users can check to see if the FIFO is empty, but at that
point there can still be up to one UART word being transmitted.

Even in the fastest modes, UART word transmission happens at about 10KHz
which is sufficiently slower than most processor/bus speeds.  As a
result, it is possible for the processor to see that the TX FIFO is
empty before the last word has transmitted and then change the UART's
clock, which corrupts that word.  The window is small, but this issue
manifests during Zephyr boot on the HiFive1 so it's not just
theoretical.

My fix is to plumb out the busy bit from the TX module to a register
that software can access, called "txdone".  I added this as a whole new
register rather than sticking it into txctrl because there's a
variable-length field at the end of txctrl and I didn't want to have a
complicated memory map.  This bit informs software that the TX queue is
empty and the last word has actually been transmitted.  In order to
inform software of the new register I've bumped the device tree's compat
string.

The semantics of the bit are intended to allow software to ensure that
no word is currently transmitting from the UART so the clock can safely
be changed before resuming transmission.  This bit is designed to allow
for two modes:

* Software wants to change the clock ASAP.  In this case TX is disabled,
  which causes the TX FIFO to stop firing.  As a result, when TX is
  disabled it doesn't matter if there are words in the FIFO so the done
  bit is just directly the busy bit.
* Software wants to change the clock when nothing is busy.  In this
  case TX is left enabled, so the TX FIFO must be empty before the clock
  can be changed.  As a result, when TX is enabled the done bit can only
  be set when the FIFO is empty.

In order to allow software to detect the presence of this new bit, I've
added the "sifive,uart1" compat string.  Since this device is backwards
compatible it retains the "sifive,uart0" string.

As usual, I've given this patch absolutely no testing: in this case I
haven't even compiled it.  I wrote most of this before 9am, so I'm sure
it's broken :)